Support custom type diff
It seems that either the type could be the same with different data, or
the entire thing is a different entry in the custom type.

I'm not sure if a simple string is enough to represent the custom type.
It might make sense to have the module & the custom type name and the
type entry name.

I'm not sure what the official term for 'type entry' is.